arc: Don't pretend we may enable user mode with kflag
The only way to enable user mode is to return to it from kernel mode
with restoring of status32 register from the context as we do in
z_new_thread().

As it is stated in Programmer's Reference Manual (PRM) kflag instruction
doesn't change status32.U flag so drop useless and misleading code.

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>