drivers: i2c: designware i2c: code refactor
in old code framework, when you access the register
of dw i2c peripheral, the compile will do some optimization
to use byte or half world memory operation, such as: lds, ldh,
sts, sth, which will cause memory access exception in hsdk
board. And it is uncontrollable, so we need to refactor
the code, to specify to use word memory operation. Now
the code works well both in hsdk and emsk.

Signed-off-by: Watson Zeng <zhiwei@synopsys.com>