samples: coop_synchronization: new cooperative thread switching sample app
A simple application that demonstrates the most basic sanity of the kernel.
Two threads (A and B) take turns printing a greeting message to the console
They use yields and semaphores to control the whose turn it is to print.
It also enforces polled uart mode to avoid serial IRQ fires triggering
unpredictable context switch. The rate is uncontrolled. This demonstrates
kernel communication only. It's a good example to test a new arch for
testing cooperative context switching without sys timer first. It's also
useful to test cooperative-only architectures which may not have any timing
or clock facilities.
I.e. this will still work when `CONFIG_SYS_CLOCK_EXISTS`=n

Signed-off-by: Alp Sayin <alpsayin@gmail.com>